explode entry points inside lib directory
