+++ lib/file-emitter.js
@@ -1,0 +1,3 @@
+const fileState = new ( require('events') ) ;
+
+module.exports = fileState ;
